+++ plugin/src/net/groboclown/idea/p4ic/v2/server/connection/ServerConnection.java
@@ -182,6 +182,15 @@ public class ServerConnection {
     }
 
 
+    public void workOffline() {
+        statusController.disconnect();
+    }
+
+
+    public void workOnline() {
+        statusController.connect();
+    }
+